Moved function and class out of the original loop and into a separate file. Bug found when user attempts to restart the bot and it is already in a channel or something to that degree.
