+++ index.js
@@ -3,14 +3,14 @@
 // npm install discord.js
 // npm i @discordjs/opus
 
+var ConvertTo1ChannelStream = require("./stream_util.js");
+console.log(ConvertTo1ChannelStream)
 
 const dotenv = require('dotenv');
 dotenv.config();
 const Discord = require('discord.js')
 const config = require('./config')
 const discordClient = new Discord.Client()
-
-console.log("\n\n\n 1 \n\n\n")
 
 discordClient.on('ready', () => {
   console.log(`Logged in as ${discordClient.user.tag}!`)
@@ -19,14 +19,11 @@ discordClient.login(config.discordApiToken)
 const googleSpeech = require('@google-cloud/speech')
 const googleSpeechClient = new googleSpeech.SpeechClient()
 
-console.log("\n\n\n 2 \n\n\n")
-
-
 // The channel the bot transcribes to 
 // var List = require("collections/list");
 var transChannel
 
-
+console.log("\n\n\n 1 \n\n\n")
 
 discordClient.on('message', async msg => {
   if (msg.content === 'inHere') {
@@ -40,21 +37,22 @@ discordClient.on('message', async msg => {
       return
     }
 
-    console.log("\n\n\n 3 \n\n\n")
-
-
     msg.reply('I\'m omw')
+
+    console.log("\n\n\n 2 \n\n\n")
+
 
     const connection = await memberVoiceChannel.join()
     const receiver = connection.receiver
-
-    console.log("\n\n\n 4 \n\n\n")
-
 
     connection.on('speaking', (user, speaking) => {
       if (!speaking) {
         return
       }
+
+      console.log("\n\n\n 3 \n\n\n")
+
+
       console.log(`I'm listening to ${user.username}`)
       const audioStream = receiver.createStream(user, { mode: 'pcm' })
       const requestConfig = {
@@ -63,7 +61,7 @@ discordClient.on('message', async msg => {
         languageCode: 'en-US'
       }
 
-      console.log("\n\n\n 5 \n\n\n")
+      console.log("\n\n\n 4 \n\n\n")
 
 
       const request = {
@@ -73,6 +71,8 @@ discordClient.on('message', async msg => {
         .streamingRecognize(request)
         .on('error', console.error)
         .on('data', response => {
+          console.log("\n\n\n reeeeeee \n\n\n")
+
           const transcription = response.results
             .map(result => result.alternatives[0].transcript)
             .join('\n')
@@ -81,34 +81,7 @@ discordClient.on('message', async msg => {
           console.log(transChannel.id)
           transChannel.send(`${user.username}: ${transcription}`)
         })
-        const { Transform } = require('stream')
-  
 
-        console.log("\n\n\n 6 \n\n\n")
-
-
-        function convertBufferTo1Channel(buffer) {
-          const convertedBuffer = Buffer.alloc(buffer.length / 2)
-          for (let i = 0; i < convertedBuffer.length / 2; i++) {
-            const uint16 = buffer.readUInt16LE(i * 4)
-            convertedBuffer.writeUInt16LE(uint16, i * 2)
-          }
-          return convertedBuffer
-        }
-  
-
-        console.log("\n\n\n 7 \n\n\n")
-
-
-        class ConvertTo1ChannelStream extends Transform {
-          constructor(source, options) {
-            super(options)
-          }
-  
-          _transform(data, encoding, next) {
-            next(null, convertBufferTo1Channel(data))
-          }
-        }
       const convertTo1ChannelStream = new ConvertTo1ChannelStream()
       audioStream.pipe(convertTo1ChannelStream).pipe(recognizeStream)
       audioStream.on('end', async () => {
@@ -125,85 +98,3 @@ discordClient.on('message', async msg => {
     console.log(transChannel.id)
   }
 })
-
-
-
-
-
-
-
-
-
-
-
-// To be removed in replacement for commands
-
-// discordClient.on('presenceUpdate', async (oldPresence, newPresence) => {
-//   // console.log('New Presence:', newPresence)
-//   const member = newPresence.member
-//   // console.log('Member: ', member, '\n\n\n\n\n') //test
-
-//   const presence = newPresence
-//   const memberVoiceChannel = member.voice.channel
-
-//   // console.log('Presence Activity: ',presence.activity, '\n\n\n\n\n')
-//   // console.log('Presence Activity Name: ',presence.activity.name, '\n\n\n\n\n')
-//   // console.log('Member Voice Channel: ',memberVoiceChannel, '\n\n\n\n\n')
-
-//   if (memberVoiceChannel == null) {
-//     return
-//   }
-//   const connection = await memberVoiceChannel.join()
-//   const receiver = connection.receiver
-
-//   connection.on('speaking', (user, speaking) => {
-//     if (!speaking) {
-//       return
-//     }
-//     console.log(`I'm listening to ${user.username}`)
-//     const audioStream = receiver.createStream(user, { mode: 'pcm' })
-//     const requestConfig = {
-//       encoding: 'LINEAR16',
-//       sampleRateHertz: 48000,
-//       languageCode: 'en-US'
-//     }
-//     const request = {
-//       config: requestConfig
-//     }
-//     const recognizeStream = googleSpeechClient
-//       .streamingRecognize(request)
-//       .on('error', console.error)
-//       .on('data', response => {
-//         const transcription = response.results
-//           .map(result => result.alternatives[0].transcript)
-//           .join('\n')
-//           .toLowerCase()
-//         console.log(`Transcription: ${transcription}`)
-//       })
-//       const { Transform } = require('stream')
-
-//       function convertBufferTo1Channel(buffer) {
-//         const convertedBuffer = Buffer.alloc(buffer.length / 2)
-//         for (let i = 0; i < convertedBuffer.length / 2; i++) {
-//           const uint16 = buffer.readUInt16LE(i * 4)
-//           convertedBuffer.writeUInt16LE(uint16, i * 2)
-//         }
-//         return convertedBuffer
-//       }
-
-//       class ConvertTo1ChannelStream extends Transform {
-//         constructor(source, options) {
-//           super(options)
-//         }
-
-//         _transform(data, encoding, next) {
-//           next(null, convertBufferTo1Channel(data))
-//         }
-//       }
-//     const convertTo1ChannelStream = new ConvertTo1ChannelStream()
-//     audioStream.pipe(convertTo1ChannelStream).pipe(recognizeStream)
-//     audioStream.on('end', async () => {
-//       console.log('audioStream end')
-//     })
-//   })
-// })
