+++ stream_util.js
@@ -1,0 +1,23 @@
+const Discord = require('discord.js')
+
+var {Transform} = require("stream")
+
+module.exports = class ConvertTo1ChannelStream extends Transform {
+    
+    constructor(source, options) {
+        super(options)
+    }
+    
+    _transform(data, encoding, next) {
+        next(null, convertBufferTo1Channel(data))
+    }   
+}
+
+function convertBufferTo1Channel(buffer) {
+    const convertedBuffer = Buffer.alloc(buffer.length / 2)
+    for (let i = 0; i < convertedBuffer.length / 2; i++) {
+        const uint16 = buffer.readUInt16LE(i * 4)
+        convertedBuffer.writeUInt16LE(uint16, i * 2)
+    }
+    return convertedBuffer
+}
